+++ plugins/hikari-pool/src/test/java/is/codion/plugin/hikari/pool/HikariConnectionPoolFactoryTest.java
@@ -21,8 +21,8 @@ public class HikariConnectionPoolFactoryTest {
   void test() throws Exception {
     long startTime = System.currentTimeMillis();
     HikariConnectionPoolFactory provider = new HikariConnectionPoolFactory();
-    ConnectionPoolWrapper pool = provider.createConnectionPoolWrapper(
-            H2DatabaseFactory.createDatabase("jdbc:h2:mem:HikariConnectionPoolProviderTest",
+    ConnectionPoolWrapper pool = provider.createConnectionPool(
+            H2DatabaseFactory.createDatabase("jdbc:h2:mem:HikariConnectionPoolFactoryTest",
                     Database.DATABASE_INIT_SCRIPTS.get()), UNIT_TEST_USER);
     pool.setCollectSnapshotStatistics(true);
     assertTrue(pool.isCollectSnapshotStatistics());